Fix KubernetesHook fail on an attribute absence (#25787)
Kubernetes SparkApplication manifest attribute "metadata:name" is not
mandatory if "metadata:generateName" is used instead. Recently there
was no functionality of removing already launched objects in. So the
case with no deleting already launched pods is ok. Later it was
implemented, but the author thought the "name" attribute is mandatory.
This is not true - "generateName" can be used instead. So we have 2
cases:
1. Attribute "name" is used. OK, we can delete the old object if found
2. Attribute "name" is not used. It is possible if "generateName" is
used. If so, we can not delete the old object. In this case we do
not know its name (k8s will generate it randomly using a name mask, but
we can not guess it for sure). So, we have nothing to do there, the old
pod will be not touched if exists

+++ airflow/providers/cncf/kubernetes/hooks/kubernetes.py
@@ -251,7 +251,6 @@ class KubernetesHook(BaseHook):
     ):
         """
         Creates custom resource definition object in Kubernetes
-
         :param group: api group
         :param version: api version
         :param plural: api plural
@@ -260,26 +259,32 @@ class KubernetesHook(BaseHook):
         """
         api = client.CustomObjectsApi(self.api_client)
         namespace = namespace or self._get_namespace() or self.DEFAULT_NAMESPACE
+
         if isinstance(body, str):
             body_dict = _load_body_to_dict(body)
         else:
             body_dict = body
-        try:
-            api.delete_namespaced_custom_object(
-                group=group,
-                version=version,
-                namespace=namespace,
-                plural=plural,
-                name=body_dict["metadata"]["name"],
-            )
-            self.log.warning("Deleted custom resource with the same name.")
-        except client.rest.ApiException:
-            self.log.info("Custom resource %s not found.", body_dict["metadata"]["name"])
+
+        # Attribute "name" is not mandatory if "generateName" is used instead
+        if "name" in body_dict["metadata"]:
+            try:
+                api.delete_namespaced_custom_object(
+                    group=group,
+                    version=version,
+                    namespace=namespace,
+                    plural=plural,
+                    name=body_dict["metadata"]["name"],
+                )
+
+                self.log.warning("Deleted SparkApplication with the same name")
+            except client.rest.ApiException:
+                self.log.info("SparkApplication %s not found", body_dict["metadata"]["name"])
 
         try:
             response = api.create_namespaced_custom_object(
                 group=group, version=version, namespace=namespace, plural=plural, body=body_dict
             )
+
             self.log.debug("Response: %s", response)
             return response
         except client.rest.ApiException as e:
